Consolidate MaaS token handler contract tests into main test file
Moves contract tests from separate maas_tokens_handler_contract_test.go into
maas_tokens_handler_test.go to follow the convention of co-locating tests with
their implementation in a single test file.

Changes:
- Move TestMaaSBFFAPIKeyResponseContract to maas_tokens_handler_test.go
- Move mockBFFClientForContract and mockFactoryForContract helpers
- Delete separate contract test file

No functional changes - all tests remain identical and continue to pass.
